refactor: 로그아웃 함수 직접 정의
